+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- `Data.squeeze`: squeezing the data object to the shape of the axes.
+
 ### Fixed
 - `interact2D`: fixed bug where use_imshow broke the sliders
 - `data.join` ensures valid `method` is selected